requirements.txt
streamlit>=1.35.0
groq>=0.9.0
python-dotenv>=1.0.1
edge-tts>=6.1.9
